Extract setting edition for preview to controller concern
