+++ app/controllers/campaign_controller.rb
@@ -1,12 +1,12 @@
 class CampaignController < ApplicationController
   include ApiRedirectable
+  include Previewable
 
   before_filter -> { set_expiry unless viewing_draft_content? }
 
   def show
     setup_content_item_and_navigation_helpers("/" + params[:slug])
     @publication = PublicationPresenter.new(artefact)
-    @edition = params[:edition]
     render locals: { full_width: true }
   end
 
@@ -22,9 +22,5 @@ private
       params[:slug],
       params[:edition]
     )
-  end
-
-  def viewing_draft_content?
-    params.include?('edition')
   end
 end
