+++ app/controllers/help_controller.rb
@@ -2,6 +2,7 @@ require "slimmer/headers"
 
 class HelpController < ApplicationController
   include ApiRedirectable
+  include Previewable
 
   before_filter -> { set_expiry unless viewing_draft_content? }
 
@@ -17,7 +18,6 @@ class HelpController < ApplicationController
   def show
     setup_content_item_and_navigation_helpers("/" + params[:slug])
     @publication = PublicationPresenter.new(artefact)
-    @edition = params[:edition]
     set_language_from_publication(@publication)
     render :show
   end
@@ -37,9 +37,5 @@ private
 
   def set_language_from_publication(publication)
     I18n.locale = publication.language if publication.language
-  end
-
-  def viewing_draft_content?
-    params.include?('edition')
   end
 end
